Isolate API loader external import rewriting
Move the external import rewriting helpers and compiled-binary CJS require shim out of the API module loader so loader.ts can focus on orchestration, build setup, and dynamic module loading. Keep loader.ts re-exporting the same helper functions to preserve existing imports and tests.

Constraint: Behavior-preserving refactor only; no new dependencies or public API changes

Rejected: Split import-map/adapter esbuild plugins in the same PR | would broaden the diff beyond one responsibility slice

Rejected: Change rewrite regexes while moving code | behavior risk for compiled-binary and npm import paths

Confidence: high

Scope-risk: narrow

Reversibility: clean

Directive: Keep import rewriting behavior covered by loader.test.ts before modifying regex or compiled-binary shim semantics

Tested: VF_DISABLE_LRU_INTERVAL=1 deno test --no-check --allow-all src/routing/api/module-loader/loader.test.ts src/routing/api/module-loader/loader-helpers.test.ts src/routing/api/module-loader/esbuild-plugin.test.ts src/routing/api/module-loader/http-validator.test.ts src/routing/api/module-loader/security-config.test.ts; deno fmt --check changed files; DENO_NO_PACKAGE_JSON=1 deno lint changed files; deno check src/routing/api/module-loader/loader.ts; deno task typecheck

Not-tested: Full deno task test suite before push; pre-push hook is expected to run it
